pyproject.toml
[project]
name = "agent-qa-lab"
version = "0.1.0"
description = "Trace-aware regression harness for tool-using Strands agents"
requires-python = ">=3.10"
keywords = [
    "ai-agents",
    "evaluation",
    "llm-testing",
    "tool-use",
    "strands-agents",
    "regression-testing",
    "agent-evaluation",
]
dependencies = [
    "strands-agents>=0.1.0",
    "pydantic>=2.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["schemas", "domain", "agents", "evaluators", "runners"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
